fix: fix issue #27 by updating App.tsx as per selection and relocate some functions
